move RunBenchmark
