+++ src/cli.rs
@@ -1,5 +1,5 @@
 mod parser;
 mod tracing;
 
-pub(crate) use parser::{Cli, Commands, ServeArgs};
+pub(crate) use parser::{Cli, Commands, RunBenchmark, ServeArgs};
 pub(crate) use tracing::init as init_tracing;
